IEvent added and TimerComponent tells the world its timer is done. TimerComponentTest updated Renamed component to components. Added interfaces package. 	modified:   core/src/com/bitdecay/game/objects/GameEntities.java 	modified:   core/src/com/bitdecay/game/objects/GameEntity.java 	modified:   core/src/com/bitdecay/game/objects/PlayerObject.java 	deleted:    core/src/com/bitdecay/game/objects/component/TimerComponent.java 	renamed:    core/src/com/bitdecay/game/objects/component/BitComponent.java -> core/src/com/bitdecay/game/objects/components/BitComponent.java 	renamed:    core/src/com/bitdecay/game/objects/component/ControllerComponent.java -> core/src/com/bitdecay/game/objects/components/ControllerComponent.java 	renamed:    core/src/com/bitdecay/game/objects/component/PhysicsComponent.java -> core/src/com/bitdecay/game/objects/components/PhysicsComponent.java 	renamed:    core/src/com/bitdecay/game/objects/component/PositionComponent.java -> core/src/com/bitdecay/game/objects/components/PositionComponent.java 	new file:   core/src/com/bitdecay/game/objects/components/TimerComponent.java 	renamed:    core/src/com/bitdecay/game/objects/component/WorldLinkComponent.java -> core/src/com/bitdecay/game/objects/components/WorldLinkComponent.java 	new file:   core/src/com/bitdecay/game/objects/interfaces/IEvent.java 	renamed:    core/src/com/bitdecay/game/objects/component/IUpdate.java -> core/src/com/bitdecay/game/objects/interfaces/IUpdate.java 	modified:   core/src/com/bitdecay/game/screens/GameScreen.java 	modified:   core/test/com/bitdecay/game/objects/GameEntityTest.java 	deleted:    core/test/com/bitdecay/game/objects/component/TimerComponentTest.java 	renamed:    core/test/com/bitdecay/game/objects/component/ControllerComponentTest.java -> core/test/com/bitdecay/game/objects/components/ControllerComponentTest.java 	renamed:    core/test/com/bitdecay/game/objects/component/PhysicsComponentTest.java -> core/test/com/bitdecay/game/objects/components/PhysicsComponentTest.java 	renamed:    core/test/com/bitdecay/game/objects/component/PositionComponentTest.java -> core/test/com/bitdecay/game/objects/components/PositionComponentTest.java 	new file:   core/test/com/bitdecay/game/objects/components/TimerComponentTest.java 	renamed:    core/test/com/bitdecay/game/objects/component/WorldLinkComponentTest.java -> core/test/com/bitdecay/game/objects/components/WorldLinkComponentTest.java
