+++ core/src/com/bitdecay/game/objects/components/ControllerComponent.java
@@ -1,7 +1,6 @@
-package com.bitdecay.game.objects.component;
+package com.bitdecay.game.objects.components;
 
 import com.bitdecay.jump.BitBody;
-import com.bitdecay.jump.control.BitBodyController;
 import com.bitdecay.jump.control.PlayerInputController;
 
 /**
