+++ core/src/com/bitdecay/game/objects/interfaces/IEvent.java
@@ -1,0 +1,8 @@
+package com.bitdecay.game.objects.interfaces;
+
+/**
+ * Created by MondayHopscotch on 8/2/2016.
+ */
+public interface IEvent {
+    boolean hasTriggered();
+}
