+++ core/src/com/bitdecay/game/objects/interfaces/IUpdate.java
@@ -1,4 +1,4 @@
-package com.bitdecay.game.objects.component;
+package com.bitdecay.game.objects.interfaces;
 
 /**
  * Created by MondayHopscotch on 7/24/2016.
